Ripple breaks if clientX xor clientY is zero
Should be checking nullness here, not falsiness, since zero is falsy.  Using double-equals so both null and undefined will work.  Yes, I was able to cause this breakage by clicking with a mouse.  Perhaps the keyboard detection above should be revisited too, in case you click on that one pixel [0,0]